refactor(aptu-coder): split shell_write.rs into heredoc_validation.rs + shell_scan.rs (#1298)
